Implemented a semi-fixed time step system for engine. Didn't fix choppy issue completely (which is caused by sleeping after each loop (what SFML does when you SetFrameLimit)), but still a better system than making things reliant on the frame time. Have some extra stuff in there (like the sleep at the end of main) for testing.

+++ Engine/include/Config.h
@@ -25,7 +25,7 @@
   /**
    * @return Application's FPS.
    */
-  static int GetFPS();
+  static float GetFPS();
 
 private:
   //@{
@@ -38,7 +38,7 @@ private:
 
   static int m_screenWidth;
   static int m_screenHeight;
-  static int m_FPS;
+  static float m_FPS;
 };
 
 #endif // CONFIG_H
